+++ app/src/main/java/com/example/lecturerecorder/recorder_activity/RecorderActivity.kt
@@ -83,7 +83,7 @@ class RecorderActivity : AppCompatActivity() {
         viewModel?.onStop()
     }
 
-    fun setTime(seconds: Long) {
+    fun setTime(seconds: Int) {
         runOnUiThread {
             textTimer.text = formatTime(seconds)
         }
