+++ app/src/main/java/com/example/lecturerecorder/utils/TiimeUtils.kt
@@ -1,12 +1,12 @@
 package com.example.lecturerecorder.utils
 
-fun formatTime(seconds: Long): String {
+fun formatTime(seconds: Int): String {
     val minutes = seconds / 60
     val hours = minutes / 60
 
     return "${formatForTime(hours)}:${formatForTime(minutes)}:${formatForTime(seconds % 60)}"
 }
 
-private fun formatForTime(value: Long): String {
+private fun formatForTime(value: Int): String {
     return if (value < 10) "0${value}" else value.toString()
 }
